backend/requirements.txt
google-cloud-texttospeech
dotenv
pypdf
google-generativeai
passlib
uvicorn 
fastapi
PyPDF2
gtts
itsdangerous
python-multipart
openai
google-cloud-speech
supabase
python-jose[cryptography]
python-multipart
